Make volume percentage based

+++ src/main/java/dev/JustRed23/jdautils/music/event/VolumeChangeEvent.java
@@ -7,12 +7,12 @@ import net.dv8tion.jda.api.entities.Guild;
  * Event fired when the volume of the music player changes
  * @param client the client responsible for handling guild interactions
  * @param guild the guild where the volume was changed
- * @param oldVolume the old volume
- * @param newVolume the new volume
+ * @param oldVolume the old volume (0-100)
+ * @param newVolume the new volume (0-100)
  */
 public record VolumeChangeEvent(
         JDA client,
         Guild guild,
-        float oldVolume,
-        float newVolume
+        int oldVolume,
+        int newVolume
 ) implements MusicEvent {}
